fixed use effect dependencies for data table
