+++ src/icons/campaign.jsx
@@ -2,8 +2,8 @@ export default ({ width, height }) => {
   return (
     <svg width={width || 24} height={height || 24} viewBox='0 0 24 24'>
       <path
-        fill-rule='evenodd'
-        clip-rule='evenodd'
+        fillRule='evenodd'
+        clipRule='evenodd'
         d='M6.86 13.832A2.87 2.87 0 0 1 4 10.97a2.87 2.87 0 0 1 2.861-2.86h2.463c0 1.883.063 3.862 0 5.722H6.86Zm2.5 1.083c.05.474.145.944.294 1.393.226.679.59 1.344 1.126 1.828a.374.374 0 0 1 .116.36l-.238 1.198a.367.367 0 0 1-.229.276.365.365 0 0 1-.36-.04c-.882-.615-1.526-1.517-1.947-2.496a8.33 8.33 0 0 1-.626-2.52H9.36Zm1.047-6.932v5.953c3.044.662 5.575 2.48 7.994 3.988.293.182.66.067.945-.12a.93.93 0 0 0 .438-.794V4.99a.929.929 0 0 0-.436-.793c-.282-.188-.628-.267-.94-.124-2.198 1.005-4.995 3.12-8.001 3.91Z'
         fill='currentColor'
       />
